+++ packages/react-sdk/src/components/CallControls/ScreenShareButton.tsx
@@ -34,10 +34,19 @@ export const ScreenShareButton = (props: ScreenShareButtonProps) => {
         messageAwaitingApproval={t('Awaiting for an approval to share screen.')}
         messageRevoked={t('You can no longer share your screen.')}
       >
-        <CompositeButton active={isSomeoneScreenSharing} caption={caption}>
+        <CompositeButton
+          active={isSomeoneScreenSharing}
+          caption={caption}
+          variant="primary"
+        >
           <IconButton
             icon={
               isSomeoneScreenSharing ? 'screen-share-on' : 'screen-share-off'
+            }
+            data-testid={
+              isSomeoneScreenSharing
+                ? 'screen-share-stop-button'
+                : 'screen-share-start-button'
             }
             title={caption || t('Share screen')}
             disabled={disableScreenShareButton}