+++ packages/react-sdk/src/components/CallControls/ToggleAudioButton.tsx
@@ -23,10 +23,18 @@ export const ToggleAudioPreviewButton = (
   const { microphone, isMute } = useMicrophoneState();
 
   return (
-    <CompositeButton Menu={Menu} active={isMute} caption={caption}>
+    <CompositeButton
+      Menu={Menu}
+      active={isMute}
+      caption={caption}
+      variant="secondary"
+    >
       <IconButton
         icon={!isMute ? 'mic' : 'mic-off'}
         title={caption || t('Mic')}
+        data-testid={
+          isMute ? 'preview-audio-unmute-button' : 'preview-audio-mute-button'
+        }
         onClick={() => microphone.toggle()}
       />
     </CompositeButton>
@@ -59,10 +67,16 @@ export const ToggleAudioPublishingButton = (
         messageAwaitingApproval={t('Awaiting for an approval to speak.')}
         messageRevoked={t('You can no longer speak.')}
       >
-        <CompositeButton Menu={Menu} active={isMute} caption={caption}>
+        <CompositeButton
+          Menu={Menu}
+          active={isMute}
+          caption={caption}
+          variant="secondary"
+        >
           <IconButton
             icon={isMute ? 'mic-off' : 'mic'}
             title={caption || t('Mic')}
+            data-testid={isMute ? 'audio-unmute-button' : 'audio-mute-button'}
             onClick={async () => {
               if (!hasPermission) {
                 await requestPermission();