+++ sample-apps/react/react-dogfood/components/ToggleFeedbackButton.tsx
@@ -5,14 +5,18 @@ import {
   IconButton,
   MenuToggle,
   MenuVisualType,
+  ToggleMenuButtonProps,
 } from '@stream-io/video-react-sdk';
 
 import { Feedback } from './Feedback/Feedback';
 
-export const ToggleMenuButton = forwardRef<HTMLButtonElement>((props, ref) => {
+export const ToggleMenuButton = forwardRef<
+  HTMLDivElement,
+  ToggleMenuButtonProps
+>((props, ref) => {
   return (
-    <CompositeButton>
-      <IconButton ref={ref} icon="feedback" />
+    <CompositeButton ref={ref} active={props.menuShown} variant="primary">
+      <IconButton icon="feedback" />
     </CompositeButton>
   );
 });