+++ sample-apps/react/react-dogfood/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './useChatClient';
 export * from './useWatchChannel';
 export * from './useKeyboardShortcuts';
 export * from './useWakeLock';
+export * from './useBreakpoints';